+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support to download Sentinel-2 products from Google Cloud storage.
 - Added support to download by text match.
-- Added `download by-geometry` and `download by-text` to Fire commands.
+- Added `download by-geometry` and `download by-title` to Fire commands.
 
 
 ## Released